Split the Gst sample into multiple
