+++ samples/common/Common.c
@@ -175,7 +175,8 @@ PVOID mediaSenderRoutine(PVOID customData)
 
     DLOGD("pSampleConfiguration->MediaType:", pSampleConfiguration->mediaType);
 
-    if (pSampleConfiguration->videoSource != NULL && (pSampleConfiguration->mediaType == SAMPLE_STREAMING_AUDIO_VIDEO || pSampleConfiguration->mediaType == SAMPLE_STREAMING_VIDEO_ONLY)) {
+    if (pSampleConfiguration->videoSource != NULL &&
+        (pSampleConfiguration->mediaType == SAMPLE_STREAMING_AUDIO_VIDEO || pSampleConfiguration->mediaType == SAMPLE_STREAMING_VIDEO_ONLY)) {
         DLOGD("Create video sender thread!");
         THREAD_CREATE(&pSampleConfiguration->videoSenderTid, pSampleConfiguration->videoSource, (PVOID) pSampleConfiguration);
     }
@@ -1708,7 +1709,8 @@ CleanUp:
     return retStatus;
 }
 
-STATUS addSendrecvVideoAndAudioTransceivers(PSampleConfiguration pSampleConfiguration, PSampleStreamingSession pSampleStreamingSession) {
+STATUS addSendrecvVideoAndAudioTransceivers(PSampleConfiguration pSampleConfiguration, PSampleStreamingSession pSampleStreamingSession)
+{
     ENTERS();
     STATUS retStatus = STATUS_SUCCESS;
     RtcRtpTransceiverInit audioRtpTransceiverInit, videoRtpTransceiverInit;
@@ -1930,3 +1932,95 @@ VOID onDataChannel(UINT64 customData, PRtcDataChannel pRtcDataChannel)
     dataChannelOnMessage(pRtcDataChannel, customData, onDataChannelMessage);
 }
 #endif
+
+STATUS addSendOnlyVideoRecvOnlyAudioTransceivers(PSampleConfiguration pSampleConfiguration, PSampleStreamingSession pSampleStreamingSession)
+{
+    ENTERS();
+    STATUS retStatus = STATUS_SUCCESS;
+    RtcRtpTransceiverInit audioRtpTransceiverInit, videoRtpTransceiverInit;
+    RtcMediaStreamTrack videoTrack = {0}, audioTrack = {0};
+
+    CHK(pSampleStreamingSession != NULL && pSampleStreamingSession != NULL, STATUS_NULL_ARG);
+
+    videoTrack.kind = MEDIA_STREAM_TRACK_KIND_VIDEO;
+    videoTrack.codec = pSampleConfiguration->videoCodec;
+    videoRtpTransceiverInit.direction = RTC_RTP_TRANSCEIVER_DIRECTION_SENDONLY;
+    STRCPY(videoTrack.streamId, "myKvsVideoStream");
+    STRCPY(videoTrack.trackId, "myVideoTrack");
+
+    CHK_STATUS(addTransceiver(pSampleStreamingSession->pPeerConnection, &videoTrack, &videoRtpTransceiverInit,
+                              &pSampleStreamingSession->pVideoRtcRtpTransceiver));
+
+    CHK_STATUS(configureTransceiverRollingBuffer(pSampleStreamingSession->pVideoRtcRtpTransceiver, &videoTrack,
+                                                 pSampleConfiguration->videoRollingBufferDurationSec,
+                                                 pSampleConfiguration->videoRollingBufferBitratebps));
+
+    CHK_STATUS(transceiverOnBandwidthEstimation(pSampleStreamingSession->pVideoRtcRtpTransceiver, (UINT64) pSampleStreamingSession,
+                                                sampleBandwidthEstimationHandler));
+
+    audioTrack.kind = MEDIA_STREAM_TRACK_KIND_AUDIO;
+    audioTrack.codec = pSampleConfiguration->audioCodec;
+    audioRtpTransceiverInit.direction = RTC_RTP_TRANSCEIVER_DIRECTION_RECVONLY;
+    STRCPY(audioTrack.streamId, "myKvsVideoStream");
+    STRCPY(audioTrack.trackId, "myAudioTrack");
+    CHK_STATUS(addTransceiver(pSampleStreamingSession->pPeerConnection, &audioTrack, &audioRtpTransceiverInit,
+                              &pSampleStreamingSession->pAudioRtcRtpTransceiver));
+
+    CHK_STATUS(configureTransceiverRollingBuffer(pSampleStreamingSession->pAudioRtcRtpTransceiver, &audioTrack,
+                                                 pSampleConfiguration->audioRollingBufferDurationSec,
+                                                 pSampleConfiguration->audioRollingBufferBitratebps));
+
+    CHK_STATUS(transceiverOnBandwidthEstimation(pSampleStreamingSession->pAudioRtcRtpTransceiver, (UINT64) pSampleStreamingSession,
+                                                sampleBandwidthEstimationHandler));
+CleanUp:
+    CHK_LOG_ERR(retStatus);
+
+    LEAVES();
+    return retStatus;
+}
+
+STATUS addSendOnlyVideoSendrecvAudioTransceivers(PSampleConfiguration pSampleConfiguration, PSampleStreamingSession pSampleStreamingSession)
+{
+    ENTERS();
+    STATUS retStatus = STATUS_SUCCESS;
+    RtcRtpTransceiverInit audioRtpTransceiverInit, videoRtpTransceiverInit;
+    RtcMediaStreamTrack videoTrack = {0}, audioTrack = {0};
+
+    CHK(pSampleStreamingSession != NULL && pSampleStreamingSession != NULL, STATUS_NULL_ARG);
+
+    videoTrack.kind = MEDIA_STREAM_TRACK_KIND_VIDEO;
+    videoTrack.codec = pSampleConfiguration->videoCodec;
+    videoRtpTransceiverInit.direction = RTC_RTP_TRANSCEIVER_DIRECTION_SENDONLY;
+    STRCPY(videoTrack.streamId, "myKvsVideoStream");
+    STRCPY(videoTrack.trackId, "myVideoTrack");
+
+    CHK_STATUS(addTransceiver(pSampleStreamingSession->pPeerConnection, &videoTrack, &videoRtpTransceiverInit,
+                              &pSampleStreamingSession->pVideoRtcRtpTransceiver));
+
+    CHK_STATUS(configureTransceiverRollingBuffer(pSampleStreamingSession->pVideoRtcRtpTransceiver, &videoTrack,
+                                                 pSampleConfiguration->videoRollingBufferDurationSec,
+                                                 pSampleConfiguration->videoRollingBufferBitratebps));
+
+    CHK_STATUS(transceiverOnBandwidthEstimation(pSampleStreamingSession->pVideoRtcRtpTransceiver, (UINT64) pSampleStreamingSession,
+                                                sampleBandwidthEstimationHandler));
+
+    audioTrack.kind = MEDIA_STREAM_TRACK_KIND_AUDIO;
+    audioTrack.codec = pSampleConfiguration->audioCodec;
+    audioRtpTransceiverInit.direction = RTC_RTP_TRANSCEIVER_DIRECTION_SENDRECV;
+    STRCPY(audioTrack.streamId, "myKvsVideoStream");
+    STRCPY(audioTrack.trackId, "myAudioTrack");
+    CHK_STATUS(addTransceiver(pSampleStreamingSession->pPeerConnection, &audioTrack, &audioRtpTransceiverInit,
+                              &pSampleStreamingSession->pAudioRtcRtpTransceiver));
+
+    CHK_STATUS(configureTransceiverRollingBuffer(pSampleStreamingSession->pAudioRtcRtpTransceiver, &audioTrack,
+                                                 pSampleConfiguration->audioRollingBufferDurationSec,
+                                                 pSampleConfiguration->audioRollingBufferBitratebps));
+
+    CHK_STATUS(transceiverOnBandwidthEstimation(pSampleStreamingSession->pAudioRtcRtpTransceiver, (UINT64) pSampleStreamingSession,
+                                                sampleBandwidthEstimationHandler));
+CleanUp:
+    CHK_LOG_ERR(retStatus);
+
+    LEAVES();
+    return retStatus;
+}
