+++ samples/p2p/kvsWebRTCClientMasterGstSample.c
@@ -1,399 +1,7 @@
 #include "../common/Samples.h"
-#include <gst/gst.h>
-#include <gst/app/gstappsink.h>
+#include "../common/GstMedia.h"
 
 extern PSampleConfiguration gSampleConfiguration;
-// #define VERBOSE
-
-GstElement* senderPipeline = NULL;
-
-GstFlowReturn on_new_sample(GstElement* sink, gpointer data, UINT64 trackid)
-{
-    GstBuffer* buffer;
-    STATUS retStatus = STATUS_SUCCESS;
-    BOOL isDroppable, delta;
-    GstFlowReturn ret = GST_FLOW_OK;
-    GstSample* sample = NULL;
-    GstMapInfo info;
-    GstSegment* segment;
-    GstClockTime buf_pts;
-    Frame frame;
-    STATUS status;
-    PSampleConfiguration pSampleConfiguration = (PSampleConfiguration) data;
-    PSampleStreamingSession pSampleStreamingSession = NULL;
-    PRtcRtpTransceiver pRtcRtpTransceiver = NULL;
-    UINT32 i;
-    guint bitrate;
-
-    CHK_ERR(pSampleConfiguration != NULL, STATUS_NULL_ARG, "NULL sample configuration");
-
-    info.data = NULL;
-    sample = gst_app_sink_pull_sample(GST_APP_SINK(sink));
-
-    buffer = gst_sample_get_buffer(sample);
-    isDroppable = GST_BUFFER_FLAG_IS_SET(buffer, GST_BUFFER_FLAG_CORRUPTED) || GST_BUFFER_FLAG_IS_SET(buffer, GST_BUFFER_FLAG_DECODE_ONLY) ||
-        (GST_BUFFER_FLAGS(buffer) == GST_BUFFER_FLAG_DISCONT) ||
-        (GST_BUFFER_FLAG_IS_SET(buffer, GST_BUFFER_FLAG_DISCONT) && GST_BUFFER_FLAG_IS_SET(buffer, GST_BUFFER_FLAG_DELTA_UNIT)) ||
-        // drop if buffer contains header only and has invalid timestamp
-        !GST_BUFFER_PTS_IS_VALID(buffer);
-
-    if (!isDroppable) {
-        delta = GST_BUFFER_FLAG_IS_SET(buffer, GST_BUFFER_FLAG_DELTA_UNIT);
-
-        frame.flags = delta ? FRAME_FLAG_NONE : FRAME_FLAG_KEY_FRAME;
-
-        // convert from segment timestamp to running time in live mode.
-        segment = gst_sample_get_segment(sample);
-        buf_pts = gst_segment_to_running_time(segment, GST_FORMAT_TIME, buffer->pts);
-        if (!GST_CLOCK_TIME_IS_VALID(buf_pts)) {
-            DLOGE("[KVS GStreamer Master] Frame contains invalid PTS dropping the frame");
-        }
-
-        if (!(gst_buffer_map(buffer, &info, GST_MAP_READ))) {
-            DLOGE("[KVS GStreamer Master] on_new_sample(): Gst buffer mapping failed");
-            goto CleanUp;
-        }
-
-        frame.trackId = trackid;
-        frame.duration = 0;
-        frame.version = FRAME_CURRENT_VERSION;
-        frame.size = (UINT32) info.size;
-        frame.frameData = (PBYTE) info.data;
-
-        MUTEX_LOCK(pSampleConfiguration->streamingSessionListReadLock);
-        for (i = 0; i < pSampleConfiguration->streamingSessionCount; ++i) {
-            pSampleStreamingSession = pSampleConfiguration->sampleStreamingSessionList[i];
-            frame.index = (UINT32) ATOMIC_INCREMENT(&pSampleStreamingSession->frameIndex);
-
-            if (trackid == DEFAULT_AUDIO_TRACK_ID) {
-                if (pSampleStreamingSession->pSampleConfiguration->enableTwcc && senderPipeline != NULL) {
-                    GstElement* encoder = gst_bin_get_by_name(GST_BIN(senderPipeline), "sampleAudioEncoder");
-                    if (encoder != NULL) {
-                        g_object_get(G_OBJECT(encoder), "bitrate", &bitrate, NULL);
-                        MUTEX_LOCK(pSampleStreamingSession->twccMetadata.updateLock);
-                        pSampleStreamingSession->twccMetadata.currentAudioBitrate = (UINT64) bitrate;
-                        if (pSampleStreamingSession->twccMetadata.newAudioBitrate != 0) {
-                            bitrate = (guint) (pSampleStreamingSession->twccMetadata.newAudioBitrate);
-                            pSampleStreamingSession->twccMetadata.newAudioBitrate = 0;
-                            g_object_set(G_OBJECT(encoder), "bitrate", bitrate, NULL);
-                        }
-                        MUTEX_UNLOCK(pSampleStreamingSession->twccMetadata.updateLock);
-                    }
-                }
-                pRtcRtpTransceiver = pSampleStreamingSession->pAudioRtcRtpTransceiver;
-                frame.presentationTs = pSampleStreamingSession->audioTimestamp;
-                frame.decodingTs = frame.presentationTs;
-                pSampleStreamingSession->audioTimestamp +=
-                    SAMPLE_AUDIO_FRAME_DURATION; // assume audio frame size is 20ms, which is default in opusenc
-            } else {
-                if (pSampleStreamingSession->pSampleConfiguration->enableTwcc && senderPipeline != NULL) {
-                    GstElement* encoder = gst_bin_get_by_name(GST_BIN(senderPipeline), "sampleVideoEncoder");
-                    if (encoder != NULL) {
-                        g_object_get(G_OBJECT(encoder), "bitrate", &bitrate, NULL);
-                        MUTEX_LOCK(pSampleStreamingSession->twccMetadata.updateLock);
-                        pSampleStreamingSession->twccMetadata.currentVideoBitrate = (UINT64) bitrate;
-                        if (pSampleStreamingSession->twccMetadata.newVideoBitrate != 0) {
-                            bitrate = (guint) (pSampleStreamingSession->twccMetadata.newVideoBitrate);
-                            pSampleStreamingSession->twccMetadata.newVideoBitrate = 0;
-                            g_object_set(G_OBJECT(encoder), "bitrate", bitrate, NULL);
-                        }
-                        MUTEX_UNLOCK(pSampleStreamingSession->twccMetadata.updateLock);
-                    }
-                }
-                pRtcRtpTransceiver = pSampleStreamingSession->pVideoRtcRtpTransceiver;
-                frame.presentationTs = pSampleStreamingSession->videoTimestamp;
-                frame.decodingTs = frame.presentationTs;
-                pSampleStreamingSession->videoTimestamp += SAMPLE_VIDEO_FRAME_DURATION; // assume video fps is 25
-            }
-            status = writeFrame(pRtcRtpTransceiver, &frame);
-            if (status != STATUS_SRTP_NOT_READY_YET && status != STATUS_SUCCESS) {
-#ifdef VERBOSE
-                DLOGE("[KVS GStreamer Master] writeFrame() failed with 0x%08x", status);
-#endif
-            } else if (status == STATUS_SUCCESS && pSampleStreamingSession->firstFrame) {
-                PROFILE_WITH_START_TIME(pSampleStreamingSession->offerReceiveTime, "Time to first frame");
-                pSampleStreamingSession->firstFrame = FALSE;
-            } else if (status == STATUS_SRTP_NOT_READY_YET) {
-                DLOGI("[KVS GStreamer Master] SRTP not ready yet, dropping frame");
-            }
-        }
-        MUTEX_UNLOCK(pSampleConfiguration->streamingSessionListReadLock);
-    }
-
-CleanUp:
-
-    if (info.data != NULL) {
-        gst_buffer_unmap(buffer, &info);
-    }
-
-    if (sample != NULL) {
-        gst_sample_unref(sample);
-    }
-
-    if (ATOMIC_LOAD_BOOL(&pSampleConfiguration->appTerminateFlag)) {
-        ret = GST_FLOW_EOS;
-    }
-
-    return ret;
-}
-
-GstFlowReturn on_new_sample_video(GstElement* sink, gpointer data)
-{
-    return on_new_sample(sink, data, DEFAULT_VIDEO_TRACK_ID);
-}
-
-GstFlowReturn on_new_sample_audio(GstElement* sink, gpointer data)
-{
-    return on_new_sample(sink, data, DEFAULT_AUDIO_TRACK_ID);
-}
-
-PVOID sendGstreamerAudioVideo(PVOID args)
-{
-    STATUS retStatus = STATUS_SUCCESS;
-    GstElement *appsinkVideo = NULL, *appsinkAudio = NULL;
-    GstBus* bus = NULL;
-    GstMessage* msg = NULL;
-    GError* gError = NULL;
-    gchar* gDebug = NULL;
-    GstStateChangeReturn stateChangeStatus = GST_STATE_CHANGE_SUCCESS;
-    PSampleConfiguration pSampleConfiguration = (PSampleConfiguration) args;
-
-    CHK_ERR(pSampleConfiguration != NULL, STATUS_NULL_ARG, "[KVS Gstreamer Master] Streaming session is NULL");
-
-    /**
-     * Use x264enc as its available on mac, pi, ubuntu and windows
-     * mac pipeline fails if resolution is not 720p
-     *
-     * For alaw
-     * audiotestsrc is-live=TRUE ! queue leaky=2 max-size-buffers=400 ! audioconvert ! audioresample !
-     * audio/x-raw, rate=8000, channels=1, format=S16LE, layout=interleaved ! alawenc ! appsink sync=TRUE emit-signals=TRUE name=appsink-audio
-     *
-     * For VP8
-     * videotestsrc is-live=TRUE ! video/x-raw,width=1280,height=720,framerate=30/1 !
-     * vp8enc error-resilient=partitions keyframe-max-dist=10 auto-alt-ref=true cpu-used=5 deadline=1 !
-     * appsink sync=TRUE emit-signals=TRUE name=appsink-video
-     *
-     *
-     * Raspberry Pi Hardware Encode Example
-     * "v4l2src device=\"/dev/video0\" ! queue ! v4l2convert ! "
-     * "video/x-raw,format=I420,width=640,height=480,framerate=30/1 ! "
-     * "v4l2h264enc ! "
-     * "h264parse ! "
-     * "video/x-h264,stream-format=byte-stream,alignment=au,width=640,height=480,framerate=30/1,profile=baseline,level=(string)4 ! "
-     * "appsink sync=TRUE emit-signals=TRUE name=appsink-video"
-     */
-
-    CHAR rtspPipeLineBuffer[RTSP_PIPELINE_MAX_CHAR_COUNT];
-
-    switch (pSampleConfiguration->mediaType) {
-        case SAMPLE_STREAMING_VIDEO_ONLY:
-            switch (pSampleConfiguration->srcType) {
-                case TEST_SOURCE: {
-                    if (pSampleConfiguration->videoCodec == RTC_CODEC_H265) {
-                        senderPipeline = gst_parse_launch("videotestsrc pattern=ball is-live=TRUE ! timeoverlay ! queue ! videoconvert ! "
-                                                          "video/x-raw,width=1280,height=720,framerate=25/1 ! queue ! "
-                                                          "x265enc speed-preset=veryfast bitrate=512 tune=zerolatency ! "
-                                                          "video/x-h265,stream-format=byte-stream,alignment=au,profile=main ! appsink sync=TRUE "
-                                                          "emit-signals=TRUE name=appsink-video",
-                                                          &gError);
-                    } else {
-                        senderPipeline = gst_parse_launch(
-                            "videotestsrc pattern=ball is-live=TRUE ! "
-                            "queue ! videoconvert ! videoscale ! video/x-raw,width=1280,height=720 ! "
-                            "clockoverlay halignment=right valignment=top time-format=\"%Y-%m-%d %H:%M:%S\" ! "
-                            "videorate ! video/x-raw,framerate=25/1 ! "
-                            "x264enc name=sampleVideoEncoder bframes=0 speed-preset=veryfast bitrate=512 byte-stream=TRUE tune=zerolatency ! "
-                            "video/x-h264,stream-format=byte-stream,alignment=au,profile=baseline ! "
-                            "appsink sync=TRUE emit-signals=TRUE name=appsink-video",
-                            &gError);
-                    }
-                    break;
-                }
-                case DEVICE_SOURCE: {
-                    senderPipeline = gst_parse_launch(
-                        "autovideosrc ! queue ! videoconvert ! video/x-raw,width=1280,height=720,framerate=25/1 ! "
-                        "x264enc name=sampleVideoEncoder bframes=0 speed-preset=veryfast bitrate=512 byte-stream=TRUE tune=zerolatency ! "
-                        "video/x-h264,stream-format=byte-stream,alignment=au,profile=baseline ! "
-                        " appsink sync=TRUE "
-                        "emit-signals=TRUE name=appsink-video",
-                        &gError);
-                    break;
-                }
-                case RTSP_SOURCE: {
-                    UINT16 stringOutcome =
-                        SNPRINTF(rtspPipeLineBuffer, RTSP_PIPELINE_MAX_CHAR_COUNT,
-                                 "uridecodebin uri=%s ! "
-                                 "videoconvert ! "
-                                 "x264enc name=sampleVideoEncoder bframes=0 speed-preset=veryfast bitrate=512 byte-stream=TRUE tune=zerolatency ! "
-                                 "video/x-h264,stream-format=byte-stream,alignment=au,profile=baseline ! queue ! "
-                                 "appsink sync=TRUE emit-signals=TRUE name=appsink-video ",
-                                 pSampleConfiguration->rtspUri);
-
-                    if (stringOutcome > RTSP_PIPELINE_MAX_CHAR_COUNT) {
-                        DLOGE("[KVS GStreamer Master] ERROR: rtsp uri entered exceeds maximum allowed length set by RTSP_PIPELINE_MAX_CHAR_COUNT");
-                        goto CleanUp;
-                    }
-                    senderPipeline = gst_parse_launch(rtspPipeLineBuffer, &gError);
-
-                    break;
-                }
-            }
-            break;
-
-        case SAMPLE_STREAMING_AUDIO_VIDEO:
-            switch (pSampleConfiguration->srcType) {
-                case TEST_SOURCE: {
-                    if (pSampleConfiguration->videoCodec == RTC_CODEC_H264_PROFILE_42E01F_LEVEL_ASYMMETRY_ALLOWED_PACKETIZATION_MODE &&
-                        pSampleConfiguration->audioCodec == RTC_CODEC_OPUS) {
-                        senderPipeline = gst_parse_launch(
-                            "videotestsrc pattern=ball is-live=TRUE ! "
-                            "queue ! videorate ! videoscale ! videoconvert ! video/x-raw,width=1280,height=720,framerate=25/1 ! "
-                            "clockoverlay halignment=right valignment=top time-format=\"%Y-%m-%d %H:%M:%S\" ! "
-                            "x264enc name=sampleVideoEncoder bframes=0 speed-preset=veryfast bitrate=512 byte-stream=TRUE tune=zerolatency ! "
-                            "video/x-h264,stream-format=byte-stream,alignment=au,profile=baseline ! "
-                            "appsink sync=TRUE emit-signals=TRUE name=appsink-video audiotestsrc wave=ticks is-live=TRUE ! "
-                            "queue leaky=2 max-size-buffers=400 ! audioconvert ! audioresample ! opusenc name=sampleAudioEncoder ! "
-                            "audio/x-opus,rate=48000,channels=2 ! appsink sync=TRUE emit-signals=TRUE name=appsink-audio",
-                            &gError);
-                    } else if (pSampleConfiguration->videoCodec == RTC_CODEC_H265 && pSampleConfiguration->audioCodec == RTC_CODEC_OPUS) {
-                        senderPipeline =
-                            gst_parse_launch("videotestsrc pattern=ball is-live=TRUE ! timeoverlay ! queue ! videoconvert ! "
-                                             "video/x-raw,width=1280,height=720,framerate=25/1 ! queue ! "
-                                             "x265enc speed-preset=veryfast bitrate=512 tune=zerolatency ! "
-                                             "video/x-h265,stream-format=byte-stream,alignment=au,profile=main ! appsink sync=TRUE "
-                                             "emit-signals=TRUE name=appsink-video audiotestsrc is-live=TRUE ! "
-                                             "queue leaky=2 max-size-buffers=400 ! audioconvert ! audioresample ! opusenc ! "
-                                             "audio/x-opus,rate=48000,channels=2 ! appsink sync=TRUE emit-signals=TRUE name=appsink-audio",
-                                             &gError);
-                    }
-                    // TODO: test and add more such combinations
-                    break;
-                }
-                case DEVICE_SOURCE: {
-                    senderPipeline = gst_parse_launch(
-                        "autovideosrc ! queue ! videoconvert ! video/x-raw,width=1280,height=720,framerate=25/1 ! "
-                        "x264enc name=sampleVideoEncoder bframes=0 speed-preset=veryfast bitrate=512 byte-stream=TRUE tune=zerolatency ! "
-                        "video/x-h264,stream-format=byte-stream,alignment=au,profile=baseline ! appsink sync=TRUE emit-signals=TRUE "
-                        "name=appsink-video autoaudiosrc ! "
-                        "queue leaky=2 max-size-buffers=400 ! audioconvert ! audioresample ! opusenc name=sampleAudioEncoder ! "
-                        "audio/x-opus,rate=48000,channels=2 ! appsink sync=TRUE emit-signals=TRUE name=appsink-audio",
-                        &gError);
-                    break;
-                }
-                case RTSP_SOURCE: {
-                    UINT16 stringOutcome =
-                        SNPRINTF(rtspPipeLineBuffer, RTSP_PIPELINE_MAX_CHAR_COUNT,
-                                 "uridecodebin uri=%s name=src ! videoconvert ! "
-                                 "x264enc name=sampleVideoEncoder bframes=0 speed-preset=veryfast bitrate=512 byte-stream=TRUE tune=zerolatency ! "
-                                 "video/x-h264,stream-format=byte-stream,alignment=au,profile=baseline ! queue ! "
-                                 "appsink sync=TRUE emit-signals=TRUE name=appsink-video "
-                                 "src. ! audioconvert ! "
-                                 "audioresample ! opusenc name=sampleAudioEncoder ! audio/x-opus,rate=48000,channels=2 ! queue ! "
-                                 "appsink sync=TRUE emit-signals=TRUE name=appsink-audio",
-                                 pSampleConfiguration->rtspUri);
-
-                    if (stringOutcome > RTSP_PIPELINE_MAX_CHAR_COUNT) {
-                        DLOGE("[KVS GStreamer Master] ERROR: rtsp uri entered exceeds maximum allowed length set by RTSP_PIPELINE_MAX_CHAR_COUNT");
-                        goto CleanUp;
-                    }
-                    senderPipeline = gst_parse_launch(rtspPipeLineBuffer, &gError);
-
-                    break;
-                }
-            }
-            break;
-    }
-
-    // When the senderPipeline is non-null and gError is non-null at the same time, it means
-    // GStreamer encountered a "recoverable parsing error", which doesn't guarantee the pipeline will
-    // actually work, it just means GStreamer could parse enough of the string to return a pipeline object.
-    if (gError != NULL) {
-        DLOGE("[KVS GStreamer Master] Pipeline error: %s", gError->message);
-        g_clear_error(&gError);
-        retStatus = STATUS_INVALID_OPERATION;
-        goto CleanUp;
-    }
-    CHK_ERR(senderPipeline != NULL, STATUS_NULL_ARG, "[KVS Gstreamer Master] Pipeline is NULL");
-
-    appsinkVideo = gst_bin_get_by_name(GST_BIN(senderPipeline), "appsink-video");
-    appsinkAudio = gst_bin_get_by_name(GST_BIN(senderPipeline), "appsink-audio");
-
-    if (!(appsinkVideo != NULL || appsinkAudio != NULL)) {
-        DLOGE("[KVS GStreamer Master] sendGstreamerAudioVideo(): cant find appsink, operation returned status code: 0x%08x", STATUS_INTERNAL_ERROR);
-        goto CleanUp;
-    }
-
-    if (appsinkVideo != NULL) {
-        g_signal_connect(appsinkVideo, "new-sample", G_CALLBACK(on_new_sample_video), (gpointer) pSampleConfiguration);
-    }
-    if (appsinkAudio != NULL) {
-        g_signal_connect(appsinkAudio, "new-sample", G_CALLBACK(on_new_sample_audio), (gpointer) pSampleConfiguration);
-    }
-    stateChangeStatus = gst_element_set_state(senderPipeline, GST_STATE_PLAYING);
-    DLOGD("[KVS GStreamer Master] State change null->playing returned status: %d", stateChangeStatus);
-    CHK_ERR(stateChangeStatus != GST_STATE_CHANGE_FAILURE, STATUS_FORMAT_ERROR, "State change to PLAYING failed!");
-
-    /* block until error or EOS */
-    bus = gst_element_get_bus(senderPipeline);
-    msg = gst_bus_timed_pop_filtered(bus, GST_CLOCK_TIME_NONE, GST_MESSAGE_ERROR | GST_MESSAGE_EOS);
-
-CleanUp:
-    if (gError != NULL) {
-        DLOGE("[KVS GStreamer Master] GStreamer error: %s", gError->message);
-        g_clear_error(&gError);
-        gError = NULL;
-    }
-
-    if (msg != NULL) {
-        switch (GST_MESSAGE_TYPE(msg)) {
-            case GST_MESSAGE_ERROR:
-                gst_message_parse_error(msg, &gError, &gDebug);
-                DLOGE("[KVS GStreamer Master] Received error from GStreamer: %s", gError->message);
-                DLOGD("[KVS GStreamer Master] Received debug from GStreamer: %s", gDebug);
-                g_clear_error(&gError);
-                g_free(gDebug);
-                gDebug = NULL;
-                break;
-            case GST_MESSAGE_EOS:
-                DLOGI("[KVS GStreamer Master] Received GStreamer End-Of-Stream");
-                break;
-            default:
-                DLOGE("[KVS GStreamer Master] Unhandled message type: %d", GST_MESSAGE_TYPE(msg));
-                break;
-        }
-        gst_message_unref(msg);
-        msg = NULL;
-    }
-
-    if (bus != NULL) {
-        gst_object_unref(bus);
-        bus = NULL;
-    }
-    if (senderPipeline != NULL) {
-        stateChangeStatus = gst_element_set_state(senderPipeline, GST_STATE_NULL);
-        if (stateChangeStatus == GST_STATE_CHANGE_FAILURE) {
-            DLOGE("[KVS GStreamer Master] State change to NULL failed!");
-        }
-        gst_object_unref(senderPipeline);
-        senderPipeline = NULL;
-    }
-    if (appsinkAudio != NULL) {
-        gst_object_unref(appsinkAudio);
-        appsinkAudio = NULL;
-    }
-    if (appsinkVideo != NULL) {
-        gst_object_unref(appsinkVideo);
-        appsinkVideo = NULL;
-    }
-
-    if (STATUS_SUCCEEDED(retStatus)) {
-        DLOGI("[KVS GStreamer Master] Media sender thread exited gracefully");
-    } else {
-        DLOGE("[KVS GStreamer Master] Media sender thread exited with status: 0x%08x", retStatus);
-    }
-
-    return (PVOID) (ULONG_PTR) retStatus;
-}
 
 INT32 main(INT32 argc, CHAR* argv[])
 {
@@ -449,9 +57,6 @@ INT32 main(INT32 argc, CHAR* argv[])
         if (STRCMP(argv[2], "video-only") == 0) {
             pSampleConfiguration->mediaType = SAMPLE_STREAMING_VIDEO_ONLY;
             DLOGI("[KVS Gstreamer Master] Streaming video only");
-        } else if (STRCMP(argv[2], "audio-video-storage") == 0) {
-            pSampleConfiguration->mediaType = SAMPLE_STREAMING_AUDIO_VIDEO;
-            DLOGI("[KVS Gstreamer Master] Streaming audio and video");
         } else if (STRCMP(argv[2], "audio-video") == 0) {
             pSampleConfiguration->mediaType = SAMPLE_STREAMING_AUDIO_VIDEO;
             DLOGI("[KVS Gstreamer Master] Streaming audio and video");
