+++ samples/webrtcstorage/kvsWebRTCStorageVideoOnlyMasterGstSample.c
@@ -1,0 +1,86 @@
+#include "../common/Samples.h"
+#include "../common/GstMedia.h"
+
+extern PSampleConfiguration gSampleConfiguration;
+
+INT32 main(INT32 argc, CHAR* argv[])
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    PSampleConfiguration pSampleConfiguration = NULL;
+    PCHAR pChannelName;
+
+    SET_INSTRUMENTED_ALLOCATORS();
+    UINT32 logLevel = setLogLevel();
+
+    signal(SIGINT, sigintHandler);
+
+#ifdef IOT_CORE_ENABLE_CREDENTIALS
+    CHK_ERR((pChannelName = argc > 1 ? argv[1] : GETENV(IOT_CORE_THING_NAME)) != NULL, STATUS_INVALID_OPERATION,
+            "AWS_IOT_CORE_THING_NAME must be set");
+#else
+    pChannelName = argc > 1 ? argv[1] : SAMPLE_CHANNEL_NAME;
+#endif
+
+    CHK_STATUS(createSampleConfiguration(pChannelName, SIGNALING_CHANNEL_ROLE_TYPE_MASTER, TRUE, TRUE, logLevel, &pSampleConfiguration));
+
+    useGstreamer(pSampleConfiguration, argc, argv);
+    parseSrcType(pSampleConfiguration, argc, argv, 2);
+
+    pSampleConfiguration->mediaType = SAMPLE_STREAMING_VIDEO_ONLY;
+    pSampleConfiguration->channelInfo.useMediaStorage = TRUE;
+    pSampleConfiguration->audioCodec = RTC_CODEC_OPUS;
+    pSampleConfiguration->videoCodec = RTC_CODEC_H264_PROFILE_42E01F_LEVEL_ASYMMETRY_ALLOWED_PACKETIZATION_MODE;
+    pSampleConfiguration->addTransceiversCallback = addSendOnlyVideoRecvOnlyAudioTransceivers;
+    pSampleConfiguration->customData = (UINT64) pSampleConfiguration;
+
+    // Initalize KVS WebRTC. This must be done before anything else, and must only be done once.
+    CHK_STATUS(initKvsWebRtc());
+    DLOGI("[KVS GStreamer Master] KVS WebRTC initialization completed successfully");
+
+    CHK_STATUS(initSignaling(pSampleConfiguration, SAMPLE_MASTER_CLIENT_ID));
+    DLOGI("[KVS GStreamer Master] Channel %s set up done ", pChannelName);
+
+    // Checking for termination
+    CHK_STATUS(sessionCleanupWait(pSampleConfiguration));
+    DLOGI("[KVS GStreamer Master] Streaming session terminated");
+
+CleanUp:
+
+    if (retStatus != STATUS_SUCCESS) {
+        DLOGE("[KVS GStreamer Master] Terminated with status code 0x%08x", retStatus);
+    }
+
+    DLOGI("[KVS GStreamer Master] Cleaning up....");
+
+    if (pSampleConfiguration != NULL) {
+        // Kick of the termination sequence
+        ATOMIC_STORE_BOOL(&pSampleConfiguration->appTerminateFlag, TRUE);
+
+        if (pSampleConfiguration->mediaSenderTid != INVALID_TID_VALUE) {
+            THREAD_JOIN(pSampleConfiguration->mediaSenderTid, NULL);
+        }
+
+        if (pSampleConfiguration->enableFileLogging) {
+            freeFileLogger();
+        }
+        retStatus = freeSignalingClient(&pSampleConfiguration->signalingClientHandle);
+        if (retStatus != STATUS_SUCCESS) {
+            DLOGE("[KVS GStreamer Master] freeSignalingClient(): operation returned status code: 0x%08x", retStatus);
+        }
+
+        retStatus = freeSampleConfiguration(&pSampleConfiguration);
+        if (retStatus != STATUS_SUCCESS) {
+            DLOGE("[KVS GStreamer Master] freeSampleConfiguration(): operation returned status code: 0x%08x", retStatus);
+        }
+    }
+    DLOGI("[KVS Gstreamer Master] Cleanup done");
+
+    RESET_INSTRUMENTED_ALLOCATORS();
+
+    // https://www.gnu.org/software/libc/manual/html_node/Exit-Status.html
+    // We can only return with 0 - 127. Some platforms treat exit code >= 128
+    // to be a success code, which might give an unintended behaviour.
+    // Some platforms also treat 1 or 0 differently, so it's better to use
+    // EXIT_FAILURE and EXIT_SUCCESS macros for portability.
+    return STATUS_FAILED(retStatus) ? EXIT_FAILURE : EXIT_SUCCESS;
+}
